% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emission.R
\name{get_emission}
\alias{get_emission}
\title{Function get_emission}
\usage{
get_emission(wide, days)
}
\arguments{
\item{wide}{Data in wide format (i.e., each day is a column). 
Data must be in extended representation. Here as state of (9) is missing and possibly dead, and 
a state of (10) is missing but not dead}

\item{days}{Names of the columns that contain the score for each day.}
}
\value{
Creates a 3 dimensional array that is "number of patients" x "number
of days" x "8 NIAID stats."
}
\description{
Creates Emission probabilities from data
}
\details{
Documentation of how this matrix is created is described in the vignette.
}
\examples{


}
\seealso{
\code{\link{impute}}
}
