% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{niaidMI-package}
\alias{niaidMI-package}
\title{niaidMI: multiple imputation of the NIAID OS score using a Markov model.}
\description{
Performs multiple imputation of the NIAID OS score by using a Markov model.
}
\author{
\itemize{
  \item *maintainer, author* Nathan Morris \email{morris_nathan@lilly.com}
  \item *author* Chaoran Hu \email{hu_chaoran@lilly.com}
}
}
