% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_param_est.R
\name{bootstrap_param_est}
\alias{bootstrap_param_est}
\title{Function bootstrap_param_est}
\usage{
bootstrap_param_est(
  wide,
  s,
  bin,
  b,
  days = paste0("D", 1:28),
  Em = get_emission(wide, days),
  tol = 1e-06,
  silent = FALSE
)
}
\arguments{
\item{wide}{Data in wide format (i.e., each day is a column). 
Data must be in extended representation. Here as state of (9) is missing and possibly dead, and 
a state of (10) is missing but not dead}

\item{s}{Numeric vecor representing the assigned strata for each patient as a number.}

\item{bin}{The assigned bin for each transition. Must be a numeric vector of length=(1-length(days)).}

\item{days}{Names of the columns that contain the score for each day.}

\item{Em}{Emission probabilities.}

\item{tol}{tolerance for relative reduction the log-likelihood to determine convergence of the Baum-Welch algorythm.}

\item{silent}{Allows silencing some messages.}
}
\value{
A list object with the following components
\describe{
  \item{fit}{Contains results of the primary model fit}
  \item{boot}{Contains relults from the bootstrap model fit.}
  \item{bin}{The input.}
  \item{days}{Formatted dataset.}
  \item{Em}{Attempted target.}
}
}
\description{
Fits a hidden markov model then bootstraps the data and refits the model.
}
\details{

}
\seealso{
\code{\link{tweights}}
}
