% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{niaidMI-package}
\alias{niaidMI-package}
\title{niaid_mi: multiple imputation of the niad score using a hidden markov model.}
\description{
niaid_mi: multiple imputation of the niad score using a hidden markov model.
}
\author{
\itemize{
  \item *maintainer, author* Nathan Morris \email{morris_nathan@lilly.com}
  \item *author* Chaoran Hu \email{hu_chaoran@lilly.com}
}
}
