% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{niaidMI-package}
\alias{niaidMI-package}
\title{niaidMI: Markov Model Multiple Imputation for NIAID OS.}
\description{
The implementation of Markov Model Multiple Imputation with the application to COVID-19 scale, NIAID OS.
}
\author{
\itemize{
  \item *maintainer, author* Nathan Morris \email{morris_nathan@lilly.com}
  \item *author* Chaoran Hu \email{hu_chaoran@lilly.com}
}
}
